% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bam_check_args}
\alias{bam_check_args}
\title{Performs the following checks:
- types:
    - logQ_hat is numeric vector
    - everything else matrix
- dimensions:
    - all matrices have same dims
    - logQ_hat has length equal to ncol of matrices}
\usage{
bam_check_args(datalist)
}
\description{
Performs the following checks:
- types:
    - logQ_hat is numeric vector
    - everything else matrix
- dimensions:
    - all matrices have same dims
    - logQ_hat has length equal to ncol of matrices
}
