% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{bam_check_rhat}
\alias{bam_check_rhat}
\title{Scope Rhat stat over all parameters}
\usage{
bam_check_rhat(fit, pars, ..., plot = TRUE)
}
\arguments{
\item{fit}{a stanfit object, as returned by bam_estimate}

\item{pars}{Which parameters to include--passed to stan_rhat}

\item{...}{other arguments passed to stan_rhat}

\item{plot}{Create a plot?}
}
\description{
Scope Rhat stat over all parameters
}
