% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postproc.R
\name{bam_convergence}
\alias{bam_convergence}
\title{MCMC diagnostics for BAM inference}
\usage{
bam_convergence()
}
\description{
Uses bayesplot functions to diagnose convergence and other health metrics for 
BAM MC.
}
