% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bam_data}
\alias{bam_data}
\title{Preprocess data for BAM estimation}
\usage{
bam_data(w, s = NULL, dA = NULL, Qhat, missing = c("omit", "impute"))
}
\arguments{
\item{w}{Matrix (or data frame) of widths: time as columns, space as rows}

\item{s}{Matrix of slopes: time as columns, space as rows}

\item{dA}{Matrix of area above base area: time as columns, space as rows}

\item{Qhat}{Vector of Q estimates. Needed to create prior on Q.}
}
\description{
Produces a bamdata object that can be passed to bam_estimate function
}
