% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{bam_priors}
\alias{bam_priors}
\title{Establish prior hyperparameters for BAM estimation}
\usage{
bam_priors(bamdata, variant = c("manning_amhg", "manning", "amhg"), ...)
}
\arguments{
\item{bamdata}{An object of class bamdata, as returned by \code{bam_data}}

\item{variant}{Which BAM variant to use. Options are "manning_amhg" (default), 
"manning", or "amhg".}

\item{...}{Optional manually set parameters. Quoted expressions are allowed,
e.g. \code{logQ_sd = "cv2sigma(0.8)"}.}
}
\description{
Produces a bampriors object that can be passed to bam_estimate function
}
