\name{bamr-package}
\alias{bamr-package}
\alias{bamr}
\docType{package}
\title{
\packageTitle{bamr}
}
\description{
\packageDescription{bamr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bamr}
\packageIndices{bamr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bamr}

Maintainer: \packageMaintainer{bamr}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }

