% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorCalc.R
\name{estimate_logA0}
\alias{estimate_logA0}
\title{Estimate AHG b exponent using bam data}
\usage{
estimate_logA0(bamdata)
}
\arguments{
\item{bamdata}{a bamdata object, created with bam_data()}
}
\description{
Estimate AHG b exponent using bam data
}
