% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{ln_sigsq}
\alias{ln_sigsq}
\title{Calculate lognormal sigma parameter based on truncated normal parameters}
\usage{
ln_sigsq(obs, err_sigma, a = 0)
}
\description{
Used to put measurement errors into original log-normal parameterization.
}
