% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_DAWG}
\alias{plot_DAWG}
\title{Plot a dataset that is formatted in space-down, time-across format}
\usage{
plot_DAWG(dawgmat)
}
\description{
Plot a dataset that is formatted in space-down, time-across format
}
