% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sacramento}
\alias{Sacramento}
\title{Remote-sensing data from the Sacramento River}
\format{A list of length 4 with the following elements:
\describe{
  \item{Sac_w}{A matrix of widths, in meters}
  \item{Sac_s}{A matrix of slopes (unitless)}
  \item{Sac_dA}{A matrix of partial cross-sectional area (m^2)}
  \item{Sac_QWBM}{A vector of discharge estimates from a water-balance model (m^3/s)}   
  
}}
\usage{
Sacramento
}
\description{
A dataset containing width, slope, partial cross-section area, and 
water-balance-model discharge for 367 consecutive days and 16 cross-sections 
in the Sacramento River, California.
}
\keyword{datasets}
