% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Sacramento_sm}
\alias{Sacramento_sm}
\title{A small version of the Sacramento dataset}
\format{A list of length 4 with the following elements:
\describe{
  \item{Sac_w_sm}{A matrix of widths, in meters}
  \item{Sac_s_sm}{A matrix of slopes (unitless)}
  \item{Sac_dA_sm}{A matrix of partial cross-sectional area (m^2)}
  \item{Sac_QWBM_sm}{A vector of discharge estimates from a water-balance model (m^3/s)}   
  \item{Sac_Qobs_sm}{A vector of observed discharge (for use in validation)}   
}}
\usage{
Sacramento_sm
}
\description{
A datatset with the same structure as Sacramento, but containing only a small subset of 
times and cross-sectoins, for fast testing and debugging of BAM code.
}
\keyword{datasets}
