% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamr-package.R
\docType{package}
\name{bamr-package}
\alias{bamr-package}
\alias{bamr}
\title{The 'bamr' package.}
\description{
Provides a high(er) level interface to Bayesian AMHG and Manning 
(BAM) discharge estimation, pre-compiled from Stan models.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
